requirements.txt
streamlit==1.45.1
snowflake-snowpark-python==1.34.0
snowflake-ml-python==1.9.0
pandas==2.1.4
snowflake-connector-python==3.16.0
plotly>=5.0.0
sendgrid==6.9.1
google-cloud-texttospeech
google-cloud-speech
st_audiorec